refactor: share cached instance-identity helper
Move ExecutableName() out of Tools.cpp's anonymous namespace into
Server (next to BoundPort) and add InstanceIdentity() -> {pid, port,
exe, vr}, so both inspect{kind:"state"} and (next) /api/health source
identity from one place and can't drift. pid/exe/vr are process-constant
and cached in statics -- only the bound port is read live -- so a
frequently polled health endpoint does no per-call GetModuleFileNameA /
path / string allocation. inspect now composes identity via update();
vr moves out of the RunAndWait snapshot (it's process-constant, same
value) so there is a single source for it.
